fixes #10 - map should always fill screen if it is large enough
this is following best known practices from other RPGs.  if map is
larger than the screen then the map will scroll with the character in
the middle.  when you approach the edge of the map the map will stop
scrolling at the edge of the screen and the character will be able to
walk to the edge of the screen.